refactor: moved embeddings to embeddings folder
