+++ packages/core/src/tests/Embedding.test.ts
@@ -1,4 +1,4 @@
-import { SimilarityType, similarity } from "../Embedding";
+import { SimilarityType, similarity } from "../embeddings";
 
 describe("similarity", () => {
   test("throws error on mismatched lengths", () => {
